Add IMAP/SMTP connection method but error with gmail email (check commit description) + rework of Account (gmail & co.) gestion with inheritance
Can't connect and crash at OtherCredentialService.cs line 19 :
await imap.AuthenticateAsync(account.UserName, account.Password);

+++ SmartmailAI.Core/Repository/AddressesRepository.cs
@@ -12,34 +12,52 @@ public class AddressesRepository(IDbContextFactory<AppDbContext_Address> factory
 {
 	private readonly IDbContextFactory<AppDbContext_Address> _factory = factory;
 
-	public async Task<List<AccountGmail>> GetAllAddressesAsync()
+	public async Task<List<AccountMailBase>> GetAllAddressesAsync()
 	{
 		using var _context = _factory.CreateDbContext();
 
-		return await _context.AccountGmail.ToListAsync();
+		return await _context.AccountMailBase.ToListAsync();
 	}
 
-	public async Task<AccountGmail?> GetAddressByEmailAsync(string email)
+	public async Task<AccountMailBase?> GetAddressByEmailAsync(string email)
 	{
 		using var _context = _factory.CreateDbContext();
 
-		return await _context.AccountGmail
+		return await _context.AccountMailBase
 			.FirstOrDefaultAsync(a => a.Email == email);
 	}
 
-	public async Task AddAddressAsync(AccountGmail accountGmail)
+	public async Task AddAddressAsync(AccountGmail? accountGmail = null, AccountOther? accountOther = null)
 	{
 		using var _context = _factory.CreateDbContext();
 
-		_context.AccountGmail.Add(accountGmail);
+		if (accountGmail != null)
+		{
+			_context.AccountGmail.Add(accountGmail);
+		}
+
+		if (accountOther != null)
+		{
+			_context.AccountOther.Add(accountOther);
+		}
+
 		await _context.SaveChangesAsync();
 	}
 
-	public async Task<bool> DeleteAddressAsync(AccountGmail accountGmail)
+	public async Task<bool> DeleteAddressAsync(AccountGmail? accountGmail = null, AccountOther? accountOther = null)
 	{
 		using var _context = _factory.CreateDbContext();
 
-		_context.AccountGmail.Remove(accountGmail);
+		if (accountGmail != null)
+		{
+			_context.AccountGmail.Remove(accountGmail);
+		}
+
+		if (accountOther != null)
+		{
+			_context.AccountOther.Remove(accountOther);
+		}
+
 		await _context.SaveChangesAsync();
 
 		return true;
